Add files via upload
Problem Statement
It is required to build a circuit with a potentiometer and 8 LED indicators. The
potentiometer is a variable resistance which enables the user to change the
voltage signal as desired.
Circuit Operation:
• When the user changes the potentiometer, the voltage reading is converted to
a 8-bit digital word to be displayed on the 8 LED indicators.
• For example, when the potentiometer is on maximum resistance (i.e. zero
input voltage), all 8 LED indicators are off. When it is on minimum resistance
(i.e. maximum voltage), all 8 LED indicators are on.
Design Requirements:
• Use Arduino Uno R3 board.
• Use Embedded C for programming.
• The 8 LED indicators are connected to ADC 0808 outputs.
• The potentiometer is connected to IN0.
• Include a reset push button in your hardware implementation.
• Do not use pre-defined ADC-related functions.