Add files via upload
Problem Statement:
It is required to build a circuit with 2 push buttons, a potentiometer and a DC motor.
The push buttons control the motor direction and the potentiometer controls the
motor speed.
Circuit Operation:
• When the user changes the potentiometer, the speed of the motor changes
accordingly. For example, when the potentiometer is on maximum resistance (i.e.
zero input voltage), motor stops. When it is on minimum resistance (i.e. maximum
voltage), motor is at full speed.
• When the user presses the 1st push button, the motor rotates in one direction.
When the user pressed the 2nd push button, the motor rotates in the opposite one.
Design Requirements:
• Use Arduino Uno R3 board.
• Use Embedded C for programming.
• The 2 push buttons are connected to PD2-PD3.
• The potentiometer is connected to ADC0.
• Include a reset push button in your hardware implementation.
• Do not use pre-defined ADC-related or PWM-related functions.